+++ Sources/VCore/Models/Property Wrappers/PublishedKeychainStorage.swift
@@ -1,17 +1,19 @@
 //
-//  KeychainStorage.swift
+//  PublishedKeychainStorage.swift
 //  VCore
 //
-//  Created by Vakhtang Kontridze on 21.07.22.
+//  Created by Vakhtang Kontridze on 22.10.24.
 //
 
 import SwiftUI
 import Combine
 
-// MARK: - Keychain Storage
-/// Property wrapper type that reflects a value from key chain and invalidates a view on a change in value in that Keychain.
+// MARK: - Published Keychain Storage
+/// Property wrapper type that reflects a value from key chain and invalidates a view on a change in value in that key chain.
 ///
-///     @KeychainStorage("AccessToken") var accessToken: String?
+///     final class SomeClass: ObservableObject {
+///         @PublishedKeychainStorage("AccessToken") var accessToken: String?
+///     }
 ///
 /// Alternately, a `KeychainServiceConfiguration` can be passed to customize queries.
 ///
@@ -19,7 +21,7 @@ import Combine
 ///         static let someCustomConfiguration: Self = ...
 ///     }
 ///
-///     @KeychainStorage("AccessToken", configuration: .someCustomConfiguration) var accessToken: String?
+///     @PublishedKeychainStorage("AccessToken", configuration: .someCustomConfiguration) var accessToken: String?
 ///
 /// Or, a reference to an instance of `KeychainService` can be used.
 ///
@@ -27,35 +29,21 @@ import Combine
 ///         static let someCustomConfiguration: KeychainService = ...
 ///     }
 ///
-///     @KeychainStorage("AccessToken", keychainService: .someCustomConfiguration) var accessToken: String?
+///     @PublishedKeychainStorage("AccessToken", keychainService: .someCustomConfiguration) var accessToken: String?
 ///
 @propertyWrapper
-public struct KeychainStorage<Value>: DynamicProperty, Sendable
+public struct PublishedKeychainStorage<Value>: DynamicProperty, Sendable
     where Value: Sendable & Codable
 {
     // MARK: Properties
     private let keychainService: KeychainService
     private let key: String
     private let defaultValue: Value
-    
-    @State private var storage: Value
 
+    @available(*, unavailable, message: "'@PublishedKeychainStorage' is only available on properties of classes")
     public var wrappedValue: Value {
-        get {
-            storage
-        }
-        nonmutating set {
-            if Self.set(keychainService, key, newValue) {
-                storage = newValue
-            }
-        }
-    }
-    
-    public var projectedValue: Binding<Value> {
-        .init(
-            get: { wrappedValue },
-            set: { wrappedValue = $0 }
-        )
+        get { fatalError() }
+        nonmutating set { fatalError() }
     }
     
     // MARK: Initializers
@@ -68,9 +56,6 @@ public struct KeychainStorage<Value>: DynamicProperty, Sendable
         self.keychainService = keychainService
         self.key = key
         self.defaultValue = defaultValue
-        
-        let initialValue: Value = Self.get(keychainService, key, defaultValue)
-        self._storage = State(wrappedValue: initialValue)
     }
 
     /// Initializes `KeychainStorage` with `Optional` value.
@@ -85,6 +70,34 @@ public struct KeychainStorage<Value>: DynamicProperty, Sendable
             key,
             keychainService: keychainService
         )
+    }
+
+    // MARK: Observable Object Support
+    public static subscript<T>(
+        _enclosingInstance instance: T,
+        wrapped wrappedKeyPath: ReferenceWritableKeyPath<T, Value>,
+        storage storageKeyPath: ReferenceWritableKeyPath<T, Self>
+    ) -> Value {
+        get {
+            let keychainService: KeychainService = instance[keyPath: storageKeyPath].keychainService
+            let key: String = instance[keyPath: storageKeyPath].key
+            let defaultValue: Value = instance[keyPath: storageKeyPath].defaultValue
+            
+            return Self.get(keychainService, key, defaultValue)
+        }
+        set {
+            let keychainService: KeychainService = instance[keyPath: storageKeyPath].keychainService
+            let key: String = instance[keyPath: storageKeyPath].key
+            
+            Self.set(keychainService, key, newValue)
+            
+            if
+                let instance = instance as? any ObservableObject,
+                let observableObjectPublisher = (instance.objectWillChange as any Publisher) as? ObservableObjectPublisher
+            {
+                observableObjectPublisher.send()
+            }
+        }
     }
     
     // MARK: Get & Set